### 2025-07-31 bfc4428
adapt utilities for autocorrelation timescales
