### 2025-07-31 bc4f03b
minor edits to run scripts for autocorr calc
