### 2025-08-12 9f4dde5
update paper files and create sphinx structure
