+++ joss/paper.bib
@@ -298,20 +298,12 @@ archivePrefix = {arXiv},
        author = {MacDonald, Ryan J. and Madhusudhan, Nikku},
         title = {POSEIDON: Multidimensional atmospheric retrieval of exoplanet spectra},
  howpublished = {Astrophysics Source Code Library, record ascl:2412.028},
-         year = 2024,
-        month = dec,
+         year = {2024},
+        month = {December},
+          DOI = {10.21105/joss.04873},
           eid = {ascl:2412.028},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024ascl.soft12028M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@misc{StellarFit,
-  author    = {Michael Radica},
-  title     = {StellarFit: Bayesian parameter estimation for stellar spectra},
-  year      = {2024},
-  publisher = {GitHub},
-  journal   = {GitHub repository},
-  url       = {https://github.com/radicamc/StellarFit}
 }
 
 @ARTICLE{bell_eureka_2022,
