+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. STCTM documentation master file, created by
+   sphinx-quickstart on Tue Aug 12 11:01:27 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to STCTM's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
