+++ docs/api/exotune_utilities.rst
@@ -1,0 +1,6 @@
+.. _exotune_utilities:
+
+Utility functions for stellar spectrum retrievals
+=================
+.. automodule:: stctm.exotune_utilities
+   :members:
