+++ docs/api/pystellspec.rst
@@ -1,0 +1,6 @@
+.. _pystellspec:
+
+Definition of stellar spectrum object
+=================
+.. automodule:: stctm.pystellspec
+   :members:
