+++ docs/api/pytransmspec.rst
@@ -1,0 +1,6 @@
+.. _pytransmspec:
+
+Definition of transmission spectrum object
+=================
+.. automodule:: stctm.pytransmspec
+   :members:
