### 2025-08-12 e1bd40e
first complete version with SPHINX (run through)
