+++ docs/api/stellar_retrieval_utilities.rst
@@ -1,6 +1,6 @@
 .. _stellar_retrieval_utilities:
 
 Utility functions for TLS retrievals
-=================
+====================================
 .. automodule:: stctm.stellar_retrieval_utilities
    :members:
