+++ docs/content/usage.rst
@@ -1,0 +1,3 @@
+STCTM Usage Guide
+=================
+
