+++ docs/api/pystellspec.rst
@@ -1,6 +1,6 @@
 .. _pystellspec:
 
 Definition of stellar spectrum object
-=================
+=====================================
 .. automodule:: stctm.pystellspec
    :members:
