+++ docs/index.rst
@@ -1,20 +1,31 @@
-.. STCTM documentation master file, created by
-   sphinx-quickstart on Tue Aug 12 11:01:27 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to STCTM's documentation!
+Welcome to STCTM's Documentation!
 =================================
+
+STCTM: the STellar ConTamination Modeling framework
+
+**stctm** performs spectral retrievals on **exoplanet transmission spectra** and (out of transit) **stellar spectra** assuming the stellar surface can be explained by a combination of spectral components.
+
+Previous Uses of the Code
+-------------------------
+
+- Lim et al., 2023 (`TRAPPIST-1 b <https://ui.adsabs.harvard.edu/abs/2023ApJ...955L..22L/abstract>`_)
+- Roy et al., 2023 (`GJ 9827 d <https://ui.adsabs.harvard.edu/abs/2023ApJ...954L..52R/abstract>`_)
+- Piaulet-Ghorayeb et al., 2024 (`GJ 9827 d <https://ui.adsabs.harvard.edu/abs/2024ApJ...974L..10P/abstract>`_)
+- Radica et al., 2025 (`TRAPPIST-1 c <https://ui.adsabs.harvard.edu/abs/2025ApJ...979L...5R/abstract>`_)
+- Ahrer et al., 2025 (`GJ 3090 b <https://ui.adsabs.harvard.edu/abs/2025arXiv250420428A/abstract>`_)
+
+
+Table of Contents
+=================
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
+   content/installation
+   content/usage
+   content/citations
+   content/contributions
+   api/api
 
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
+*Have fun!*
