+++ docs/content/citations.rst
@@ -1,0 +1,41 @@
+Citations
+=========
+
+If you use this code in your work, please (pending entry for the JOSS submission) use the following BibTeX entries for citation:
+
+::
+
+    @misc{piaulet_stctm_2024,
+        author       = {Caroline Piaulet-Ghorayeb},
+        title        = {{stctm: Stellar contamination retrievals and modeling for small planet transmission spectra}},
+        month        = aug,
+        year         = 2024,
+        doi          = {10.5281/zenodo.13153251},
+        version      = {1.0.0},
+        publisher    = {Zenodo},
+        url          = {https://doi.org/10.5281/zenodo.13153251}
+    }
+
+::
+
+    @ARTICLE{piaulet-ghorayeb_gj9827_2024,
+        author = {{Piaulet-Ghorayeb}, Caroline and {Benneke}, Bj{\"o}rn and {Radica}, Michael and {Raul}, Eshan and {Coulombe}, Louis-Philippe and {Ahrer}, Eva-Maria and {Kubyshkina}, Daria and {Howard}, Ward S. and {Krissansen-Totton}, Joshua and {MacDonald}, Ryan J. and {Roy}, Pierre-Alexis and {Louca}, Amy and {Christie}, Duncan and {Fournier-Tondreau}, Marylou and {Allart}, Romain and {Miguel}, Yamila and {Schlichting}, Hilke E. and {Welbanks}, Luis and {Cadieux}, Charles and {Dorn}, Caroline and {Evans-Soma}, Thomas M. and {Fortney}, Jonathan J. and {Pierrehumbert}, Raymond and {Lafreni{\`e}re}, David and {Acu{\~n}a}, Lorena and {Komacek}, Thaddeus and {Innes}, Hamish and {Beatty}, Thomas G. and {Cloutier}, Ryan and {Doyon}, Ren{\'e} and {Gagnebin}, Anna and {Gapp}, Cyril and {Knutson}, Heather A.},
+        title = "{JWST/NIRISS Reveals the Water-rich ``Steam World'' Atmosphere of GJ 9827 d}",
+        journal = {\apjl},
+        keywords = {Exoplanet atmospheres, Exoplanet atmospheric composition, Exoplanet atmospheric evolution, Exoplanet structure, Planetary atmospheres, Exoplanet astronomy, 487, 2021, 2308, 495, 1244, 486, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+        year = 2024,
+        month = oct,
+        volume = {974},
+        number = {1},
+        eid = {L10},
+        pages = {L10},
+        doi = {10.3847/2041-8213/ad6f00}
+    }
+
+Additional Citations
+====================
+
+If you use MSG for your stellar models, please also cite their JOSS paper
+`https://doi.org/10.21105/joss.04 <https://doi.org/10.21105/joss.04>`_.
+
+In any case, please cite where you obtained the stellar models from.
