### 2025-08-17 979251a
update for v1 of full documentation
