+++ stctm/stellar_retrieval_utilities.py
@@ -159,6 +159,8 @@ def init_default_and_fitted_param(Tphot, met, logg_phot,
                                   Dscale_guess=7000, logg_het_guess=None):
     """
 
+    Initialize default and fitted parameters.
+
     Parameters
     ----------
     Tphot : float
@@ -187,6 +189,7 @@ def init_default_and_fitted_param(Tphot, met, logg_phot,
         Initial guess for Dscale parameter. Default is 7000.
     logg_het_guess : float
         Initial guess for logg_het. Default is None.
+
     Returns
     -------
     dict of default param values and list of fitted parameters.
@@ -252,12 +255,15 @@ def init_default_and_fitted_param(Tphot, met, logg_phot,
 def init_default_and_fitted_param_fromDict(input_para, Dscale_guess=7000):
     """
 
+    Use an input dictionary to initialize default and fitted parameters
+
     Parameters
     ----------
     input_para : dict
         dictionary of input parameters
     Dscale_guess : float
         Initial guess for Dscale parameter. Default is 7000.
+
     Returns
     -------
     dict of default param values and list of fitted parameters.
@@ -363,6 +369,7 @@ def parse_gaussian_inputs(input_para):
 def get_derived_param(param):
     """
     Get derived model parameters from the fitted parameters
+
     Parameters
     ----------
     param : dict
@@ -386,14 +393,15 @@ def get_default_logg(params):
     """
     Get the default values for the stellar and heterogeneity log g
 
-    ----------
+
     Parameters
     ----------
     params: dict
     Dictionary obtained when reading in the ini file
 
+    Returns
     -------
-    This dictionary is updated with 2 more entries added by this function
+    None (dictionary updated in place with 2 more entries)
     """
 
     # logg for MCMC
@@ -415,6 +423,7 @@ def get_param_priors(param, gaussparanames, hyperp_gausspriors=[],
                      fitLogfSpotFac=[False, False], hyperp_logpriors=[],
                      T_grid=[2300, 10000], logg_grid=[], Dscale_guess=7000.):
     """
+    Query for parameter prior values.
 
     Parameters
     ----------
@@ -544,6 +553,8 @@ def make_wavelength_array(wv_min_um=0.2, wv_max_um=6.0, resPower=1000, use_pymsg
 def load_phoenix_model(Teff, M, logG, wv_target=None, wv_edges=None, resPower_target=None,
                        wv_min_um=0.2, wv_max_um=6.0, use_pymsg=True, pymsg_specgrid=None):
     """
+    Load a PHOENIX model given stellar model parameters
+
     Parameters
     ----------
     Teff : float
@@ -566,6 +577,7 @@ def load_phoenix_model(Teff, M, logG, wv_target=None, wv_edges=None, resPower_ta
         whether or not to use pymsg to load the model
     pymsg_specgrid : msg.SpecGrid() object
         pymsg grid already loaded from file name
+
     Returns
     -------
     wavePhoenix : array
@@ -621,7 +633,6 @@ def load_phoenix_model(Teff, M, logG, wv_target=None, wv_edges=None, resPower_ta
 
 def get_stellar_model_grid(input_para, preprocessed=False):
     """
-
     Load a stellar model grid based on user-defined input parameters.
 
 
@@ -631,6 +642,7 @@ def get_stellar_model_grid(input_para, preprocessed=False):
         dictionary of input parameters
     preprocessed: bool
         whether or not the parameter ranges already have the right format
+
     Returns
     -------
     tuple
[…]
