+++ docs/api/exotune_utilities.rst
@@ -4,3 +4,4 @@ Utility functions for stellar spectrum retrievals
 =================================================
 .. automodule:: stctm.exotune_utilities
    :members:
+   :no-index:

[…]
