+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 import os
 
 def read_version():
@@ -9,6 +9,19 @@ def read_version():
                 return line.split(delim)[1]
 
 setup(
-    # other args...
+    name="stctm",
     version=read_version(),
+    packages=find_packages(),
+    include_package_data=True,
+    install_requires=[
+        # put your dependencies here, e.g.:
+        # "numpy", "matplotlib", "astropy", "emcee",
+    ],
+    entry_points={
+        "console_scripts": [
+            "stctm_TLS=stctm.cli.stctm_tls:main",
+            # later you’ll add exotune too:
+            # "stctm_exotune=stctm.cli.stctm_exotune:main",
+        ],
+    },
 )
