+++ pyproject.toml
@@ -1,3 +1,6 @@
 [build-system]
 requires = ["setuptools>=64.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project.scripts]
+stctm_TLS = "stctm.cli.stctm_tls:main"
