### 2025-08-17 4b78531
CLI implemented in TLS, exotune, and docs
