+++ stctm/runners/exotune_runner.py
@@ -1,0 +1,480 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Oct 31 12:40:05 2019
+
+@author: caroline
+
+Fit a stellar spectrum using combination of stellar models
+v4: with ini file for inputs
+v4: implement custom pool for multiprocessing
+v3: implement log priors
+v2: implement error inflation
+v1: implement parallelization
+CLEAN_v0: move functions to utilities scripts
+
+"""
+## ------------------- Import modules ------------------#
+import os
+
+# make sure your environment variables are set up (see example below)
+# os.environ['CRDS_SERVER_URL'] = "https://jwst-crds.stsci.edu"
+# os.environ['CRDS_PATH'] = "/home/caroline/crds_cache"
+# os.environ['PYSYN_CDBS'] = "/home/caroline/trds"
+
+import numpy as np
+import matplotlib.pyplot as plt
+import pdb
+os.environ["OMP_NUM_THREADS"] = "1"
+
+import stctm.pystellspec as psspec
+import stctm.exotune_utilities as xtu
+import stctm.stellar_retrieval_utilities as sru
+import emcee
+import astropy.table as table
+import sys
+import astropy.io as aio
+from matplotlib.gridspec import GridSpec
+from pathlib import Path
+
+## The main code starts here
+
+def main(argv):
+    '''
+        Example:
+     	python exotune_runscript_v5_clean_20250422.py template_ini_exotune.ini
+        '''
+
+    if len(argv) > 1:
+        iniFile = argv[1]
+    else:
+        iniFile = 'template_ini_exotune.ini'
+    if not os.path.exists(iniFile):
+        raise FileNotFoundError('USER ERROR: iniFile does not exist.')
+
+    params, label_run = xtu.parse_all_input_exotune_iniFile(iniFile)
+    print("\nStarting Run:", label_run)
+
+    ini_path = Path(iniFile).resolve()
+    ini_dir = ini_path.parent
+    results_root = (ini_dir / "../../exotune_results").resolve()
+
+    # Folder for the early "optimize_param" exit
+    res_suffix = params["res_suffix"]  # used later
+    pre_runname = f"preprocessOnly{label_run}_{res_suffix}"
+    pre_results_folder = results_root / pre_runname
+    pre_results_folder.mkdir(parents=True, exist_ok=True)
+
+    ## Pre-processing
+    if params["start_from_timeseries"]:
+        #  Reading in the observed spectrum
+        timeUnique, waveUnique, optspec2D = xtu.read_in_2d_spec(params["path_to_stellar_spec_ts"])
+
+        # Remove time intervals from 2D array to calculate median
+        fig1, timeUnique, optspec2D = xtu.calc_and_plot_timeMasked_mednorm_lc(timeUnique, optspec2D, params)
+
+        print("\nCalculating median out-of-transit spectrum and use to initialize StellSpec object...")
+        spec = xtu.get_StellSpec_from_spec_timeseries(waveUnique,optspec2D, params, flux_calibrated=True)
+
+    else:
+        print("\nRead in median out-of-transit spectrum and use to initialize StellSpec object...")
+        t = aio.ascii.read(params["path_to_spec"]) # read in the spectrum that was provided
+        spec = psspec.StellSpec(t, inputtype=params["spec_format"])
+
+    ##  Apply operations to the StellSpec object
+
+    spec, fig2 = xtu.prep_spec_waveMasked(spec, params)
+
+    print("\nObtain values from spec object to save as picklable object for MCMC...")
+    # get the values we need from the spec object while keeping it picklable (only numpy arrays)
+    waveMin = np.array(spec.waveMin)
+    waveMax = np.array(spec.waveMax)
+    yval = np.array(spec.yval)
+    yerrUpp = np.array(spec.yerrUpp)
[…]
