+++ stctm/cli/stctm_tls.py
@@ -1,13 +1,9 @@
 # stctm/cli/stctm_tls.py
-from __future__ import annotations
 import argparse
 import sys
 from pathlib import Path
 
-def main(argv: list[str] | None = None) -> int:
-    """
-    Console entry point for:  stctm_TLS your_ini_file.ini
-    """
+def main(argv=None) -> int:
     argv = sys.argv[1:] if argv is None else argv
 
     p = argparse.ArgumentParser(
@@ -15,13 +11,14 @@ def main(argv: list[str] | None = None) -> int:
         description="Run TLS stellar contamination retrievals using an INI config."
     )
     p.add_argument("ini_path", help="Path to your INI file")
-    args = p.parse_args(argv)
+
+    # IMPORTANT: don't error on extra -key=value overrides
+    args, unknown = p.parse_known_args(argv)
 
     ini = Path(args.ini_path)
     if not ini.is_file():
         p.error(f"INI file not found: {ini}")
 
-
+    # Hand argv (including overrides) through to the runner
     from stctm.runners.tls_runner import main as runner_main
-    # The original runner expects argv with the ini path at index 1
-    return int(runner_main(["stctm_TLS", str(ini)]) or 0)
+    return int(runner_main(["stctm_TLS", str(ini), *unknown]) or 0)
