+++ stctm/runners/tls_runner.py
@@ -38,6 +38,7 @@ import stctm.stellar_retrieval_utilities as sru
 import stctm.exotune_utilities as xtu
 
 import matplotlib.gridspec as gridspec
+from pathlib import Path
 
 
 ## The main code starts here
@@ -100,29 +101,41 @@ def main(argv):
 
         runname = "fit"
         for p in fitparanames:
-            runname = runname + "_"+p
-        runname = runname + "_"+ params["res_suffix"]
-        results_folder = "../../stellar_contamination_results/"+runname+"/"
-        print("Results folder:", results_folder)
+            runname = runname + "_" + p
+        runname = runname + "_" + params["res_suffix"]
 
-        # Setup for post-processing
+        # --- NEW: base everything off the INI file location ---
+        ini_path = Path(iniFile).resolve()
+        ini_dir = ini_path.parent
 
-        if not os.path.isdir(results_folder):
-           os.makedirs(results_folder)
+        # ../../stellar_contamination_results/ starting from
+        # stellar_contamination_analysis/your_analysis_path/
+        results_root = (ini_dir / "../../stellar_contamination_results").resolve()
 
-        this_dir = os.getcwd() + "/"
-        res_dir = os.sep.join(__file__.split(os.sep)[:-2]) + "/../stellar_contamination_results/" + runname + "/"
-        utils_script = str(sru.__file__)
+        results_folder = (results_root / runname)
+        print("Results folder:", str(results_folder))
 
-        this_script = __file__.split(os.sep)[-1]
+        results_folder.mkdir(parents=True, exist_ok=True)
+        res_dir = str(results_folder) + os.sep
 
-        sru.save_ref_files(this_dir, this_script, iniFile, utils_script, res_dir)
+        # give save_ref_files full paths
+        this_script_path = Path(__file__).resolve()
+        utils_script_path = Path(sru.__file__).resolve()
+
+        # Hand full paths to save_ref_files
+        this_dir = ""
+        this_script = str(this_script_path)  #
+        iniFile_fp = str(Path(iniFile).resolve())  #
+        utils_script = str(utils_script_path)  # full path to utilities module
+
+        # Now call exactly once
+        sru.save_ref_files(this_dir, this_script, iniFile_fp, utils_script, res_dir)
 
         print("Plotting fitted spectrum...")
 
         # plot spectrum
         fig, ax = spec.plot()
-        fig.savefig(results_folder + "stctm_fitted_spectrum.pdf")
+        fig.savefig(str(results_folder / "stctm_fitted_spectrum.pdf"))
 
         # define parameters for emcee run
         print("Setting up MCMC...")
@@ -191,12 +204,14 @@ def main(argv):
     print("\n** Fitparanames:", fitparanames)
     print("\n** Param: ", param)
 
+    results_folder = str(results_root / runname) + os.sep
+
     if 1: # chainplot
 
 
         fig, axes = sru.chainplot(sampler.chain[:, burnin:, :], labels=fitparanames)
         if params["save_fit"]:
-            fig.savefig(results_folder + "stctm_chainplot_noburnin_"+runname+'.png')
+            fig.savefig(results_folder+ "stctm_chainplot_noburnin_"+runname+'.png')
 
         fig, axes = sru.chainplot(sampler.chain[:, :, :], labels=fitparanames)
         if params["save_fit"]:
@@ -235,12 +250,13 @@ def main(argv):
[…]
