### 2025-08-22 a937293
remove duplicated functions: instead master/call
