+++ joss/paper.md
@@ -24,17 +24,18 @@ bibliography: paper.bib
 
 # Summary
 
-Transmission spectroscopy is the most promising method for the near-term study of small-planet atmospheres and the search for atmospheres on temperate rocky worlds.
-At the same time, the small planets that are most amenable for such atmospheric probes orbit small M dwarf stars, which challenge transmission spectroscopy studies via "stellar contamination", or the transit light source effect (TLSE).
-Specifically, the TLSE describes the fact that spectral contrasts between bright and dark spots at the stellar surface can leave wavelength-dependent imprints in transmission spectra that may be mistaken for planetary atmosphere absorption.  
+Transmission spectroscopy is a key avenue for the near-term study of small-planet atmospheres and the most promising method when it comes to searching for atmospheres on temperate rocky worlds, which are often too cold for planetary emission to be detectable.
+At the same time, the small planets that are most amenable for such atmospheric probes orbit small M dwarf stars. This "M-dwarf opportunity" has encountered a major challenge because of late-type stars' magnetic activity, which lead to the formation of spots and faculae at their surface. 
+If inhomogeneously distributed throughout the photosphere, this phenomenon can give rise to "stellar contamination", or the transit light source effect (TLSE).
+Specifically, the TLSE describes the fact that spectral contrasts between bright and dark spots at the stellar surface outside of the transit chord can leave wavelength-dependent imprints in transmission spectra that may be mistaken for planetary atmosphere absorption.  
 
 As the field becomes increasingly ambitious in the search for signs of even thin atmospheres on small exoplanets, the TLSE is becoming a limiting factor, and it becomes imperative to develop robust inference methods to disentangle planetary and stellar contributions to the observed spectra.
 Here, I present `stctm`, a flexible Bayesian retrieval framework to model the impact of the TLSE on any exoplanet transmission spectrum, and infer the range of stellar surface parameters that are compatible with the observations in the absence of any planetary contribution. 
-With the `exotune` sub-module, users can also perform retrievals directly on out-of-transit stellar spectra in order to constrain the extent to which the TLSE can impact any planet's transmission spectrum.
-The input data formats, stellar models, and fitted parameters are easily tunable to the needs of the user and the code is fully parallelized to enable fast inferences. 
+With the `exotune` sub-module, users can also perform retrievals directly on out-of-transit stellar spectra in order to place data-driven priors on the extent to which the TLSE can impact any planet's transmission spectrum.
+The input data formats, stellar models, and fitted parameters are easily tunable using human-readable files and the code is fully parallelized to enable fast inferences. 
 
 # Statement of need
-The interpretation of high-precision exoplanet transmission spectra facilities like the Hubble Space Telescope (HST) and the James Webb Space Telescope (JWST) is increasingly dependent on a robust accounting for the effects of stellar contamination, particularly for small planets orbiting small stars. Despite a growing awareness of the Transit Light Source Effect (TLSE; [@Rackham:2018; @TRAPPIST1_JWST_2024]), the community currently lacks flexible, open-source tools that allow for robust modeling and retrieval of stellar contamination signatures.
[…]
