+++ joss/paper.bib
@@ -124,6 +124,20 @@ archivePrefix = {arXiv},
   Note      = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@Article{PiauletGhorayeb:2025,
+ Author     =  {Piaulet-Ghorayeb, Caroline and Benneke, Bj{\"o}rn and Turbet, Martin and Moore, Keavin and Roy, Pierre-Alexis and Lim, Olivia and Doyon, Ren{\'e} and Fauchez, Thomas J. and Albert, Lo{\"i}c and Radica, Michael and Coulombe, Louis-Philippe and Lafreni{\`e}re, David and Cowan, Nicolas B. and Belzile, Danika and Musfirat, Kamrul and Kaur, Mehramat and L'Heureux, Alexandrine and Johnstone, Doug and MacDonald, Ryan J. and Allart, Romain and Dang, Lisa and Kaltenegger, Lisa and Pelletier, Stefan and Rowe, Jason F. and Taylor, Jake and Turner, Jake D.},
+  Title     =  {Strict limits on potential secondary atmospheres on the temperate rocky exo-Earth TRAPPIST-1 d},
+  Journal   = {arXiv e-prints},
+  Keywords  = {Earth and Planetary Astrophysics, Solar and Stellar Astrophysics},
+  Year      = {2025},
+  Month     = {Aug},
+  Eprint    = {2508.08416},
+  ArchivePrefix = {arXiv},
+  PrimaryClass  = {astro-ph.EP},
+  URL       = {https://ui.adsabs.harvard.edu/abs/2024ApJ...974L..10P},
+  Note      = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @Article{Radica:2025,
   Author    = {Radica, Michael and Piaulet-Ghorayeb, Caroline and Taylor, Jake and Coulombe, Louis-Philippe and Benneke, Bj{\"o}rn and Albert, Loic and Artigau, {\'E}tienne and Cowan, Nicolas B. and Doyon, Ren{\'e} and Lafreni{\`e}re, David and L'Heureux, Alexandrine and Lim, Olivia},
   Title     = {Promise and Peril: Stellar Contamination and Strict Limits on the Atmosphere Composition of TRAPPIST-1 c from JWST NIRISS Transmission Spectra},
