+++ stctm/exotune_utilities.py
@@ -36,66 +36,24 @@ import shutil
 
 import configparser
 import argparse
+from . import stellar_retrieval_utilities as sru
 
 ## Parsing
 
-def parse_range_string(s, as_type=float):
-    """
-    Convert a string representing two numeric values into a list.
-
-    Parameters
-    ----------
-    s : str
-        A string with two numeric values separated by an underscore (e.g., "1.4_5.4").
-    as_type : type, optional
-        The type to cast the values to, either `float` (default) or `int`.
-
-    Returns
-    -------
-    list
-        A list containing the two values, cast to the specified type.
-
-    """
-    return [as_type(x) for x in s.split('_')]
 
 def parse_gaussian_inputs(input_para):
     """
-    Parse Gaussian parameter names and corresponding hyperparameter priors.
+    Forwarder to the master implementation in stellar_retrieval_utilities.
 
     Parameters
     ----------
     input_para : dict
-        dictionary of input parameters
 
     Returns
     -------
-    tuple
-        - gaussparanames : np.ndarray
-            Array of parameter names.
-        - hyperp_gausspriors : list of lists
-            List of [mean, std] values for each parameter.
+    (gaussparanames: np.ndarray, hyperp_gausspriors: list[list[float]])
     """
-
-    # Handle parameter names
-    if input_para["gaussparanames"] == "":
-        gaussparanames = np.array([])
-    else:
-        gaussparanames = np.array(input_para["gaussparanames"].split('_'))
-
-    # Handle hyperparameter priors
-    if input_para["hyperp_gausspriors"] == "":
-        hyperp_gausspriors = []
-    else:
-        hyperp_gausspriors = [parse_range_string(pair) for pair in input_para["hyperp_gausspriors"].split('|')]
-
-    if len(hyperp_gausspriors) != len(gaussparanames):
-        raise ValueError(f"Number of hyperparameter ranges ({len(hyperp_gausspriors)}) does not match "
-                         f"number of Gaussian parameters ({len(gaussparanames)}).")
-
-    if len(gaussparanames):
-        print("** Using Gaussian prior on ", list(gaussparanames))
-
-    return gaussparanames, hyperp_gausspriors
+    return sru.parse_gaussian_inputs(input_para)
 
 def parse_all_input_exotune_iniFile(iniFile):
     """
@@ -273,26 +231,26 @@ def parse_all_input_exotune_iniFile(iniFile):
         if params[para_to_parse] == "":
             params[para_to_parse] = []
         else:
-            params[para_to_parse] = [parse_range_string(pair) for pair in params[para_to_parse].split('|')]
+            params[para_to_parse] = [sru.parse_range_string(pair) for pair in params[para_to_parse].split('|')]
 
     # Stellar models grid
-    params["logg_range"] = parse_range_string(params["logg_range"])
+    params["logg_range"] = sru.parse_range_string(params["logg_range"])
 
     if params["Teff_range"] == "default":
-        params["Teff_range"] = [np.min([2300. - params["Teffstar"], -100.]) + params["Teffstar"],
+        params["Teff_range"] = [np.max([params["Teffstar"]-1000., 2300.]),
                                 params["Teffstar"] + 1000.]
     else:
-        params["Teff_range"] = parse_range_string(params["Teff_range"])
[…]

[…]
