+++ stctm/stellar_retrieval_utilities.py
@@ -28,6 +28,10 @@ import configparser
 import argparse
 import shutil
 import os
+try:
+    import tomllib as tomli  # Py 3.11+
+except Exception:
+    import tomli             # backwards-compatible
 
 ##  --- Constants --- #
 
@@ -37,10 +41,37 @@ h = const.h.value
 
 ##  --- Processing inputs --- #
 
+def config_get(cfg, table, key, default=None, required=True):
+    """
+    Helper to fetch nested TOML keys.
+
+    Parameters
+    ----------
+    cfg: dict of parameter values
+    table: subsection in the TOML file
+    key: key to fetch
+    default: default value
+    required: boolean: if True, raise error if the entry is missing.
+
+    Returns
+    -------
+    Fetched value for the key in the cfg file
+    """
+    if table not in cfg:
+        if required:
+            raise KeyError(f"Missing [{table}] table in config.")
+        return default
+    t = cfg[table]
+    if key not in t:
+        if required:
+            raise KeyError(f"Missing key '{key}' in [{table}]")
+        return default
+    return t[key]
 
 def parse_all_input_iniFile(iniFile):
     """
-    Parse an INI configuration file and command-line style arguments into a parameter dictionary.
+    Parse a TOML (new, preferred) file,
+    or a legacy INI configuration file and command-line style arguments into a parameter dictionary.
 
     Parameters
     ----------
@@ -54,100 +85,99 @@ def parse_all_input_iniFile(iniFile):
         This dictionary can be used directly in downstream modeling and fitting code (after a bit of further parsing).
     """
 
-    config = configparser.ConfigParser()
-    config.read(iniFile)
 
-    # %% Inputs for fit
+    # Detect format by extension; prefer TOML
+    ext = (iniFile.split(".")[-1].lower() if "." in iniFile else "")
+    use_toml = ext in {"toml", "tml", "tom"}
 
-    print('\nReading in the inputs...')
+    if use_toml:
+        with open(iniFile, "rb") as f:
+            cfg = tomli.load(f)
+    else:
+        # Legacy INI path
+        cfp = configparser.ConfigParser()
+        cfp.read(iniFile)
+        cfg = {
+            "paths_and_labels": dict(cfp.items("paths and labels")) if cfp.has_section("paths and labels") else {},
+            "stellar_params": dict(cfp.items("stellar params")) if cfp.has_section("stellar params") else {},
+            "stellar_models": dict(cfp.items("stellar models")) if cfp.has_section("stellar models") else {},
+            "MCMC_params": dict(cfp.items("MCMC params")) if cfp.has_section("MCMC params") else {},
+            "priors": dict(cfp.items("priors")) if cfp.has_section("priors") else {},
+            "plotting": dict(cfp.items("plotting")) if cfp.has_section("plotting") else {},
+        }
+        # Cast booleans
+        if "save_fit" in cfg["paths_and_labels"]:
+            cfg["paths_and_labels"]["save_fit"] = cfp.getboolean("paths and labels", "save_fit")
+        for k in (
+        "parallel", "fitspot", "fitfac", "fitThet", "fitTphot", "fitlogg_phot", "fitlogg_het", "fitDscale"):
+            if k in cfg["MCMC_params"]:
+                cfg["MCMC_params"][k] = cfp.getboolean("MCMC params", k)
 
+    # Build argparse using config defaults to enable CLI overrides
     parser = argparse.ArgumentParser(description='Inputs for stellar contamination retrieval.')
 
-    # --- [paths and labels] ---
-    parser.add_argument('-path_to_spec', help='Path to spectrum file',
-                        default=config.get('paths and labels', 'path_to_spec'))
[…]
