+++ stctm/exotune_utilities.py
@@ -37,8 +37,367 @@ import shutil
 import configparser
 import argparse
 from . import stellar_retrieval_utilities as sru
+try:
+    import tomllib as tomli  # Py 3.11+
+except Exception:
+    import tomli             # backwards-compatible
 
 ## Parsing
+
+def convert_to_bool(x):
+    """
+    Convert a value into a boolean.
+
+    Parameters
+    ----------
+    x : Any
+        Value to convert. Can be bool, str, int, or None.
+
+    Returns
+    -------
+    bool
+        Converted boolean value.
+    """
+    if isinstance(x, bool):
+        return x
+    if x is None:
+        return False
+    s = str(x).strip().lower()
+    if s in ("true", "t", "1", "yes", "y"):
+        return True
+    if s in ("false", "f", "0", "no", "n"):
+        return False
+    return bool(x)
+
+
+def parse_range_pair(value, as_type=float):
+    """
+    Normalize a 2-value range into a list [min, max].
+
+    Parameters
+    ----------
+    value : str, list, tuple, or None
+        The range to parse. Can be "a_b", [a, b], (a, b), or None/"".
+
+    as_type : type, optional
+        The type to cast values to (default: float).
+
+    Returns
+    -------
+    list
+        A 2-element list [min, max] of the given type, or [] if empty.
+    """
+    if value is None or value == "":
+        return []
+    if isinstance(value, (list, tuple)):
+        if len(value) != 2:
+            raise ValueError(f"Range must have 2 items, got {value}")
+        return [as_type(value[0]), as_type(value[1])]
+    parts = str(value).split("_")
+    if len(parts) != 2:
+        raise ValueError(f"Could not parse range '{value}'. Expected 'min_max' or [min,max].")
+    return [as_type(parts[0]), as_type(parts[1])]
+
+# table aliases defined for transition from .ini file to .toml file as input
+_TABLE_ALIASES = {
+    "choice of inputs": ("choice of inputs", "choice_of_inputs"),
+    "preprocessing": ("preprocessing",),
+    "saving options": ("saving options", "saving_options"),
+    "stellar params": ("stellar params", "stellar_params"),
+    "stellar models": ("stellar models", "stellar_models"),
+    "MCMC params": ("MCMC params", "MCMC_params"),
+    "priors": ("priors",),
+    "plotting": ("plotting",),
+}
+
+def parse_mask_ranges(v):
+    """
+    Normalize a mask specification into a list of ranges.
+
+    Parameters
+    ----------
+    v : str, list, or None
+        Mask specification. Can be:
+        - "" or None -> []
+        - "a_b|c_d"  -> [[a, b], [c, d]]
+        - [[a, b], [c, d]] -> returned as list of floats
+
+    Returns
+    -------
+    list
+        A list of [min, max] float pairs.
+    """
+    if v in (None, ""):
+        return []
+    if isinstance(v, str):
+        out = []
+        for pair in v.split("|"):
+            pair = pair.strip()
+            if not pair:
+                continue
+            out.append(parse_range_pair(pair, float))
+        return out
+    return [[float(p[0]), float(p[1])] for p in v]
+
+
+def preprocess_gaussians(params):
+    """
+    Pre-normalize Gaussian prior fields so the sru.parse_gaussian_inputs() forwarder
+    receives the shapes it expects.
+
+    Parameters
+    ----------
+    params : dict
+        Parameter dictionary to be mutated in place.
+
+    Returns
+    -------
+    None
+        Mutates 'gaussparanames' and 'hyperp_gausspriors' in params.
+    """
+    # gaussparanames: "", None -> []; "A_B" -> ["A","B"]; list/array -> list
+    names = params.get("gaussparanames")
+    if names in (None, ""):
[…]
