+++ docs/content/running_exotune.rst
@@ -37,17 +37,17 @@ For reference, here is how ``StellSpec`` objects are initialized from a table in
 * You may read your data using any compatible astropy table format (see the `Astropy Table I/O documentation <https://docs.astropy.org/en/stable/table/io.html>`_).
 * Ensure your wavelength array, ``waveMin``, and ``waveMax`` are in microns (convert if necessary).
 
-The recommended approach is to provide your spectrum file in one of these formats and then instantiate your object by specifying the format (via the ``spec_format`` parameter in your ``.ini`` file, see below).
+The recommended approach is to provide your spectrum file in one of these formats and then instantiate your object by specifying the format (via the ``spec_format`` parameter in your ``.toml`` file, see below).
 
 In summary, as long as your input file fits one of these supported formats or you extend ``StellSpec`` for your format, you will be able to analyze your observed stellar spectrum with *exotune*.
 
 Setting up an *exotune* retrieval: Run instructions
 ---------------------------------------------------
 
-The ``.ini`` file structure largely mirrors that of TLS retrievals (see :ref:`runningTLS` for folder layout). The key exotune-specific modifications are:
+The ``.toml`` file structure largely mirrors that of TLS retrievals (see :ref:`runningTLS` for folder layout). The key exotune-specific modifications are:
 
 * The analysis folder is ``exotune_analysis/any_analysis_folder_name/``. Create a new folder for each project under ``exotune_analysis/``, e.g., ``template_exotune_analysis/``.
-* The default INI file is ``template_ini_exotune.ini`` which you can use as a starting point to build your own.
+* The default TOML file is ``template_ini_exotune.toml`` which you can use as a starting point to build your own.
 * Results are saved in ``exotune_results/``, a sibling folder to ``exotune_analysis/``. Each run produces a subfolder in ``exotune_results/``.
 * Ensure all environment variables and paths are correctly set (refer to comments at the top of the run script for requirements), copied here::
 
@@ -58,40 +58,42 @@ The ``.ini`` file structure largely mirrors that of TLS retrievals (see :ref:`ru
 
 Example command to run exotune using the CLI helper, after navigating to your analysis folder::
 
-    stctm_exotune template_ini_exotune.ini
+    stctm_exotune template_ini_exotune.toml
 
-Inputs can be edited in the INI file or passed as command-line arguments. For instance, to switch off ``fitspot``::
+Inputs can be edited in the TOML file or passed as command-line arguments. For instance, to switch off ``fitspot``::
 
-    stctm_exotune template_ini_exotune.ini -fitspot=0
+    stctm_exotune template_ini_exotune.toml -fitspot=0
 
-Modifying the INI file to make it your own
+Modifying the TOML file to make it your own
 ------------------------------------------
 
-The sections and parameters below correspond directly to the exotune ``.ini`` file.
+The sections and parameters below correspond directly to the exotune ``.toml`` file.
 
 Choosing inputs and starting format
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
[…]
