### 2025-08-23 03495a3
add extra docs for RTD
