+++ stctm/__init__.py
@@ -1,2 +1,2 @@
 # stctm/__init__.py
-__version__ = "2.0.2"
+__version__ = "2.1.0"
