+++ stctm/__init__.py
@@ -1,2 +1,2 @@
 # stctm/__init__.py
-__version__ = "2.1.0"
+__version__ = "2.1.1"
