### 2025-08-25 979b566
add results from paper generation
