### 2025-08-25 9435e95
small paper header fixes
