+++ joss/paper.tex
@@ -278,11 +278,11 @@ spectra. \textit{Journal of Open Source Software}, ¿VOL?(¿ISSUE?), ¿PAGE?. \u
 \title{STCTM: a forward modeling and retrieval framework for stellar
 contamination and stellar spectra}
 
-        \author[1,2]{Caroline Piaulet-Ghorayeb}
+        \author[1, 2]{Caroline Piaulet-Ghorayeb}
     
       \affil[1]{Department of Astronomy \& Astrophysics, University of
 Chicago, 5640 South Ellis Avenue, Chicago, IL 60637, USA}
-      \affil[2]{E. Margaret Burbridge Prize Postdoctoral Fellow}
+      \affil[2]{E. Margaret Burbidge Prize Postdoctoral Fellow}
   \date{\vspace{-7ex}}
 
 \begin{document}
