### 2025-08-25 ef50cd8
minor updates to paper text
