### 2025-08-25 ba346e4
add latest paper files
