### 2025-12-05 #8269 @MartianColonist
Hi @Romain-Thomas-Shef,

My PhD student has been using STCTM, so I'd be happy to review this JOSS submission!
