### 2026-01-17 #8269 @warrickball
Hi everyone — For clarity I wanted to post here about our recently [announced new changes to our requirements](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future), which now appear in our docs on [what the paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html), and [scope definition](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance). Because @Romain-Thomas-Shef started editing this submission before 2026, it will be reviewed under the previous scope, and we've now labelled the submission appropriately.

Some of the automatic posts in the review process might refer to the new, updated scope and therefore not be relevant. We're working on automating this but it might involve manual intervention after the review starts. If you have any questions, just ask me or @Romain-Thomas-Shef.
