### 2026-01-18 #9859 @editorialbot
[REVIEW]: STCTM: a forward modeling and retrieval framework for stellar contamination and stellar spectra
**Submitting author:** @cpiaulet (<a href="http://orcid.org/0000-0002-2875-917X">Caroline Piaulet-Ghorayeb</a>)
**Repository:** https://github.com/cpiaulet/stctm
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.1
