### 2026-01-18 #9859 @Romain-Thomas-Shef
Dear @MartianColonist, @nenasedk

Welcome to the review thread of this paper! All communication will happen here from now on.

Please update the checklist and give feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html

We will aim to get the review done in the next few weeks. As you know, JOSS reviews are a bit different from other journal in the sense that we are trying to be a bit more interactive. Do not hesitate to open issues (and even PRs) on the software repo and to update this issue with your feedback even if you have not finished your review.

Finally, if you have any questions, please do not hesitate to ping me (@Romain-Thomas-Shef )
