### 2026-03-23 #9859 @nenasedk
Hi @cpiaulet, just wanted to check in and see if you will restructure the text of the paper. In the meantime, I have a few comments on using STCTM.

1) Is it possible to put together a python interface for the retrieval? I think it would be useful to be able to run the retrieval in a python script and immediately be able to post-process and plot additional results.

2) I think it would be convenient if STCTM could take in a retrieval name and automatically generate the necessary folders, so that they are consistently defined each time and the user doesn't need to worry about setting them up ahead of time (changing from the process [here](https://stctm.readthedocs.io/en/latest/content/running_TLS.html#run-instructions).

3) It would also be nice if the package was directly installable with pip/uv/conda, rather than needing to clone the repository. From the JOSS guidelines, "Ideally [package dependencies] should be handled with an automated package management solution."

4) In the [installation process](https://stctm.readthedocs.io/en/latest/content/installation.html#testing-your-installation), you mention that the CRDS environment variables need to be set. Please link to the space telescope CRDS documentation so that the user can see how to properly install and configure CRDS.

5) Both the paper and the documentation mention exotune, and go to significant lengths detailing _how_ to run an exotune retrieval, but it is not made clear _why_ the user should run an exotune retrieval rather than a TLS retrieval. Please include a brief blurb in the documentation explaining the differences between the methods, and when the user should use one or the other.
