### 2026-03-24 #9859 @Romain-Thomas-Shef
Test for @cpiaulet
