### 2026-04-02 #9859 @cpiaulet
> Hi [@Romain-Thomas-Shef](https://github.com/Romain-Thomas-Shef), I'm starting to push through the review now, sorry for the delay!
>
> [@cpiaulet](https://github.com/cpiaulet) I have a few questions about the paper structure: in the checklist there are several sections that are not explicitly included in the paper by name, but are mostly covered in the text. For example, the state of the field is discussed in the statement of need, with references to other similar software in the similar tools section. Could you please restructure the paper to follow the required JOSS section of Summary, Statement of need, State of the field, Software design, Research impact and AI usage disclosure? Thanks!

Sounds good, I'll put a new version of the paper together by the end of next week.
