### 2026-04-08 #9859 @Romain-Thomas-Shef
@MartianColonist Did you have the time to look at the paper?
